+++ docs/source/conf.py
@@ -118,7 +118,7 @@ html_sidebars = {
 
 # Everything intersphinx's to Python.
 intersphinx_mapping = {
-    'python': ('https://docs.python.org/3.5', None),
+    'python': ('https://docs.python.org/3.6', None),
 }
 
 # Autodoc settings