+++ CHANGELOG.rst
@@ -1,2 +1,7 @@
 Changelog
 =========
+
+Release 0.0.1 (Development)
+---------------------------
+* Create initial implementation of :class:`picklepipe.PicklePipe`.
+* Create :meth:`picklepipe.make_pipe_pair` to create a connected pair of :class:`picklepipe.PicklePipe` instances.